Carousel starting to work
Learned a lot today that helped get the Carousel off the ground and
running.

It is now rotating as intended locally and on Bluehost. I still have
some questions about how to best manage files, and how to best point it
to a local or external CSS, but it’s working for the time being.

The caduceus8.png is a file icon when it is opened as a tab in a
desktop web browser.

Now to start customizing this for HS and have pages point to html/php
files for data entry, “about”, etc.