Creation of getstarted.html
This was simply an effort to knock out another page. In this instance
it was simply the addition of savisky-model-1.php code (I know it’s
outdated) and to show those contents in the midst of our carousel-style
CSS. Again, I’m centering using the Bootstrap box model, and I don’t
know if that’s right, but it’s my best guess for now.

I did a ton of arranging and indenting code to make everything line up
properly, and while everything is not exactly centered, I believe I am
getting a working understanding of how to make sections act as we’d
like them to within our CSS.

Again, very much a placeholder, but another step in the right direction.

I would assume we would want a username/password wall before this, or a
“tour”, like we’ve discussed.

Also, I moved the icon into the pics folder and updated each page’s
.html to reflect.