Creation of clients.html, learnmore.html, and contact.html
BOO YAH. THE TRAIN ROLLS ON.

Added contact info to the footer of every page.

Again, placeholder pages. Hopefully the “clients” page fills out with
more testimonials, videos, pics, etc.

“About” and “learn more” are similar. Eventually, I would like to
differentiate: “about” to be more about the company (history, purpose,
mission, etc). I would like “learn more” to be the place where people
stop to gather more intel before they pull the trigger on “get started”
and actually use the model. I would like “learn more” to be as robust
as possible. Right now, it’s a polished turd of marketing language I
simply made up just now. It’s actually not a bad start, but it’s mostly
some lines of thought for arguments in which we’d want to present
information for. I have no sacred cows here. I have a plan in mind for
the layout that would be easier to discuss in person. I plan on trying
to code this.

“Contact” is simply the bottom section of the “about” page.